# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  admin.CollectCategoryInput:
    properties:
      category_id:
        type: integer
      external_category_id:
        type: integer
    required:
    - category_id
    - external_category_id
    type: object
  admin.CollectCategoryMapItem:
    properties:
      category_id:
        type: integer
      external_category_id:
        type: integer
      id:
        type: integer
      source_id:
        type: integer
    type: object
  admin.CollectLogItem:
    properties:
      collect_count:
        type: integer
      created_at:
        type: string
      duration_sec:
        type: integer
      id:
        type: integer
      source_id:
        type: integer
      source_name:
        type: string
      status:
        type: integer
    type: object
  admin.CollectNowRequest:
    properties:
      data_range:
        enum:
        - today
        - last1d
        - last3d
        - last1w
        - last1m
        - all
        type: string
    required:
    - data_range
    type: object
  admin.CollectSourceItem:
    properties:
      collect_url:
        type: string
      cron_expr:
        type: string
      data_range:
        type: string
      id:
        type: integer
      last_collect_at:
        type: string
      name:
        type: string
      play_source_id:
        type: integer
      play_source_name:
        type: string
      schedule_enabled:
        type: integer
      status:
        type: integer
      type:
        type: integer
    type: object
...
  admin.CreateCollectSourceRequest:
    properties:
      api_key:
        maxLength: 255
        type: string
      collect_url:
        maxLength: 500
        minLength: 1
        type: string
      cron_expr:
        maxLength: 100
        type: string
      data_range:
        maxLength: 20
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      play_source_id:
        type: integer
      type:
        enum:
        - 1
        - 2
        type: integer
    required:
    - collect_url
    - name
    - play_source_id
    - type
    type: object
...
  admin.RemoteCategoryItem:
    properties:
      type_id:
        type: integer
      type_name:
        type: string
      type_pid:
        type: integer
    type: object
  admin.RemoteCategoryResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/admin.RemoteCategoryItem'
        type: array
    type: object
...
  admin.SetCollectCategoriesRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/admin.CollectCategoryInput'
        type: array
    type: object
...
  admin.UpdateCollectSourceRequest:
    properties:
      api_key:
        maxLength: 255
        type: string
      collect_url:
        maxLength: 500
        minLength: 1
        type: string
      cron_expr:
        maxLength: 100
        type: string
      data_range:
        maxLength: 20
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      play_source_id:
        type: integer
      type:
        enum:
        - 1
        - 2
        type: integer
    type: object
...
  response.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      cause:
        type: string
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 小橘TV - 影视站系统
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 小橘TV
  version: "1.0"
paths:
...
  /api/admin/v1/collect-sources:
    get:
      consumes:
      - application/json
      description: 获取采集源分页列表
      parameters:
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.CollectSourceItem'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 采集源列表
      tags:
      - 管理端｜采集管理
    post:
      consumes:
      - application/json
      description: 创建一个新的采集源
      parameters:
      - description: 采集源参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CreateCollectSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.CollectSourceItem'
              type: object
      security:
      - BearerAuth: []
      summary: 新建采集源
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定采集源
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除采集源
      tags:
      - 管理端｜采集管理
    put:
      consumes:
      - application/json
      description: 更新指定采集源信息
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 采集源参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateCollectSourceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.CollectSourceItem'
              type: object
      security:
      - BearerAuth: []
      summary: 更新采集源
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}/categories:
    get:
      consumes:
      - application/json
      description: 获取指定采集源的分类映射列表
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/admin.CollectCategoryMapItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 采集源分类映射列表
      tags:
      - 管理端｜采集管理
    post:
      consumes:
      - application/json
      description: 替换指定采集源的分类映射
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类映射参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.SetCollectCategoriesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/admin.CollectCategoryMapItem'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 设置采集源分类映射
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}/collect:
    post:
      consumes:
      - application/json
      description: 立即执行一次采集任务，返回 data.started=true
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      - description: 立即采集参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.CollectNowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 立即采集
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}/remote-categories:
    get:
      consumes:
      - application/json
      description: 从远程采集源拉取分类列表（仅苹果CMS）
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.RemoteCategoryResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取远程分类
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}/schedule/disable:
    post:
      consumes:
      - application/json
      description: 禁用指定采集源的定时采集，返回 data.disabled=true
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 禁用定时采集
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect-sources/{id}/schedule/enable:
    post:
      consumes:
      - application/json
      description: 启用指定采集源的定时采集（需已绑定分类和设置cron表达式），返回 data.enabled=true
      parameters:
      - description: 采集源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties:
                    type: boolean
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: 启用定时采集
      tags:
      - 管理端｜采集管理
  /api/admin/v1/collect/logs:
    get:
      consumes:
      - application/json
      description: 获取采集日志分页列表
      parameters:
      - in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - in: query
        maximum: 1000000
        minimum: 1
        name: page
        type: integer
      - in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - in: query
        name: source_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageData'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/admin.CollectLogItem'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 采集日志列表
      tags:
      - 管理端｜采集管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"